+++ src/pic/simulator/specialfunctionregisters/Trisb.java
@@ -11,7 +11,7 @@ import pic.simulator.pins.Pin;
 public class Trisb extends SpecialFunctionRegister
 {
     private final HashMap<Integer, IOPin> pins;
-    private byte                          value;
+    private short                          value;
 
     public Trisb(PicProcessor processor)
     {
@@ -33,7 +33,7 @@ public class Trisb extends SpecialFunctionRegister
     @Override
     public void setValue(short value)
     {
-        this.value = (byte) (value & 0x1f);
+        this.value = (short) (value & 0x1f);
 
         if ((value & 0b00000001) > 0)
             pins.get(Pin.RB0).setToInput();
@@ -85,7 +85,7 @@ public class Trisb extends SpecialFunctionRegister
     @Override
     public void reset()
     {
-        value = (byte)0xFF;
+        value = 0xFF;
 
         for (IOPin pin : pins.values())
             pin.setToOutput();
